+++ lib/ex_check/check.ex
@@ -1,24 +1,25 @@
 defmodule ExCheck.Check do
   @moduledoc false
 
-  alias ExCheck.{Command, Config, GraphExecutor, Printer, Project}
+  alias ExCheck.{Command, Config, Printer}
+  alias __MODULE__.{Compiler, Pipeline}
 
   def run(opts) do
-    config = Config.load()
-    tools_from_config = Config.get_tools(config)
-    opts_from_config = Config.get_opts(config)
-    opts_merged = Keyword.merge(opts_from_config, opts)
+    {tools, config_opts} = Config.load()
+    opts = Keyword.merge(config_opts, opts)
 
-    compile_and_run_tools(tools_from_config, opts_merged)
+    compile_and_run_tools(tools, opts)
   end
 
   defp compile_and_run_tools(tools, opts) do
+    {compiler, others} = Compiler.compile(tools, opts)
+
     start_time = DateTime.utc_now()
-    compiler = run_compiler(tools, opts)
-    finished = if run_others?(compiler), do: run_others(tools, opts), else: []
+    compiler_result = run_compiler(compiler)
+    others_results = if run_others?(compiler_result), do: run_others(others, opts), else: []
     total_duration = DateTime.diff(DateTime.utc_now(), start_time)
 
-    all_results = [compiler | finished]
+    all_results = [compiler_result | others_results]
     failed_results = Enum.filter(all_results, &match?({:error, _, _}, &1))
 
     reprint_errors(failed_results)
@@ -26,11 +27,7 @@ defmodule ExCheck.Check do
     maybe_set_exit_status(failed_results)
   end
 
-  defp run_compiler(tools, opts) do
-    compiler = List.keyfind(tools, :compiler, 0) || raise("compiler tool definition missing")
-    compiler = prepare_tool(compiler, opts)
-    compiler = with {:disabled, _} <- compiler, do: {:pending, {:compiler, "mix compile", []}}
-
+  defp run_compiler(compiler) do
     run_tool(compiler)
   end
 
@@ -41,97 +38,31 @@ defmodule ExCheck.Check do
   end
 
   defp run_others(tools, opts) do
-    {pending, skipped} =
-      tools
-      |> List.keydelete(:compiler, 0)
-      |> Enum.sort_by(&get_tool_order/1)
-      |> Enum.map(&prepare_tool(&1, opts))
-      |> Enum.reject(&match?({:disabled, _}, &1))
-      |> Enum.split_with(&match?({:pending, _}, &1))
-
+    {pending, skipped} = Enum.split_with(tools, &match?({:pending, _}, &1))
     {finished, broken} = run_tools(pending, opts)
 
-    new_skipped =
-      Enum.map(broken, fn {name, [unresolved_dep | _], _} ->
-        {:skipped, name, ["broken tool dependency ", :bright, to_string(unresolved_dep), :normal]}
-      end)
-
-    finished ++ skipped ++ new_skipped
+    finished ++ skipped ++ broken
   end
 
-  # To better understand what's happening here and what exactly `GraphExecutor` does consider that
-  # if we wouldn't care about tool dependencies we could use following `run_tools` implementation:
-  #
-  #     defp run_tools(tools, opts) do
-  #       if Keyword.get(opts, :parallel, true) do
-  #         {tools |> Enum.map(&start_tool/1) |> Enum.map(&await_tool/1), []}
-  #       else
-  #         {Enum.map(tools, &run_tool/1), []}
-  #       end
-  #     end
-  #
-  # Refer to `GraphExecutor` code comments for more info.
   defp run_tools(tools, opts) do
-    parallel = Keyword.get(opts, :parallel, true)
-
     tool_deps =
       Enum.map(tools, fn tool = {:pending, {name, _, opts}} ->
         deps = Keyword.get(opts, :run_after, [])
         {name, deps, tool}
       end)
 
-    GraphExecutor.run(
-      tool_deps,
-      parallel: parallel,
-      start_fn: &start_tool/1,
-      collect_fn: &await_tool/1
-    )
+    {finished, broken} =
+      Pipeline.run(
+        tool_deps,
+        throttle_fn: &throttle_tools(&1, &2, opts),
+        start_fn: &start_tool/1,
+        collect_fn: &await_tool/1
+      )
+
+    broken = for {name, [dep | _], _} <- broken, do: {:skipped, name, {:run_after, dep}}
+
+    {finished, broken}
   end
-
-  defp prepare_tool({name, config}, opts) do
-    command = Keyword.fetch!(config, :command)
-    command_opts = Keyword.take(config, [:cd, :env, :enable_ansi, :run_after])
-    require_deps = Keyword.get(config, :require_deps, [])
-    require_files = Keyword.get(config, :require_files, [])
-
-    cond do
-      Keyword.get(config, :enabled, true) == false ->
-        {:disabled, name}
-
-      Keyword.has_key?(opts, :only) && !Enum.any?(opts, &(&1 == {:only, name})) ->
-        {:disabled, name}
-
-      Enum.any?(opts, fn i -> i == {:except, name} end) ->
-        {:disabled, name}
-
-      missing_dep = find_missing_dep(require_deps) ->
-        {:skipped, name, ["missing package ", :bright, to_string(missing_dep), :normal]}
-
-      missing_file = find_missing_file(require_files) ->
-        {:skipped, name, ["missing file ", :bright, missing_file, :normal]}
-
-      true ->
-        {:pending, {name, command, command_opts}}
-    end
-  end
-
-  defp find_missing_dep(require_deps) do
-    Enum.find(require_deps, &(not Project.has_dep?(&1)))
-  end
-
-  defp find_missing_file(require_files) do
-    dirs = Project.get_mix_child_dirs()
-
-    Enum.find(require_files, fn file ->
-      not Enum.any?(dirs, fn dir ->
-        dir
-        |> Path.join(file)
-        |> File.exists?()
-      end)
-    end)
-  end
-
-  defp get_tool_order({_, opts}), do: Keyword.get(opts, :order, 0)
 
   defp run_tool(tool) do
     tool
@@ -139,57 +70,50 @@ defmodule ExCheck.Check do
     |> await_tool()
   end
 
-  @ansi_code_regex ~r/(\x1b\[[0-9;]*m)/
+  defp throttle_tools(selected, running, opts) do
+    parallel = Keyword.get(opts, :parallel, true)
+
+    selected
+    |> throttle_parallel(running, parallel)
+    |> throttle_umbrella_parallel(running)
+  end
+
+  defp throttle_parallel(selected, _, true), do: selected
+  defp throttle_parallel([first_selected | _], [], false), do: [first_selected]
+  defp throttle_parallel(_, _, false), do: []
+
+  defp throttle_umbrella_parallel(selected, running) do
+    running_names = Enum.map(running, &elem(&1, 0))
+
+    Enum.reduce(selected, [], fn next = {name, _, {:pending, {_, _, opts}}}, approved ->
+      approved_names = Enum.map(approved, &elem(&1, 0))
+
+      if opts[:umbrella_parallel] == false &&
+           (includes_umbrella_instance_from_same_app?(running_names, name) ||
+              includes_umbrella_instance_from_same_app?(approved_names, name)) do
+        approved
+      else
+        approved ++ [next]
+      end
+    end)
+  end
+
+  defp includes_umbrella_instance_from_same_app?(names, match_name) do
+    Enum.any?(names, &umbrella_instance_from_same_app?(&1, match_name))
+  end
+
+  defp umbrella_instance_from_same_app?({name, _}, {name, _}), do: true
+  defp umbrella_instance_from_same_app?(_, _), do: false
 
   defp start_tool({:pending, {name, cmd, opts}}) do
-    stream = fn out ->
-      out
-      |> String.replace(@ansi_code_regex, "\\1" <> IO.ANSI.faint())
-      |> IO.write()
-    end
-
-    env = Keyword.get(opts, :env, %{})
-    final_cmd = prepare_tool_cmd(cmd, opts)
-    final_opts = Keyword.merge(opts, stream: stream, env: env, silenced: true)
-    task = Command.async(final_cmd, final_opts)
+    opts = Keyword.merge(opts, stream: true, silenced: true, tint: IO.ANSI.faint())
+    task = Command.async(cmd, opts)
 
     {:running, {name, cmd, opts}, task}
   end
 
-  defp prepare_tool_cmd(cmd, opts) when is_binary(cmd) do
-    cmd
-    |> String.split(" ")
-    |> prepare_tool_cmd(opts)
-  end
-
-  defp prepare_tool_cmd(cmd, opts) do
-    if Keyword.get(opts, :enable_ansi, true) do
-      supports_erl_config = Version.match?(System.version(), ">= 1.9.0")
-      enable_ansi(cmd, supports_erl_config)
-    else
-      cmd
-    end
-  end
-
-  # Elixir commands executed by `mix check` are not run in a TTY and will by default not print ANSI
-  # characters in their output - which means no colors, no bold etc. This makes the tool output
-  # (e.g. assertion diffs from ex_unit) less useful. We explicitly enable ANSI to fix that.
-  defp enable_ansi(["mix" | arg], true),
-    do: ["elixir", "--erl-config", erl_cfg_path(), "-S", "mix" | arg]
-
-  defp enable_ansi(["elixir" | arg], true),
-    do: ["elixir", "--erl-config", erl_cfg_path() | arg]
-
-  defp enable_ansi(["mix" | arg], false),
-    do: ["elixir", "-e", "Application.put_env(:elixir, :ansi_enabled, true)", "-S", "mix" | arg]
-
-  defp enable_ansi(cmd, _),
-    do: cmd
-
-  defp erl_cfg_path, do: Application.app_dir(:ex_check, ~w[priv enable_ansi enable_ansi.config])
-
   defp await_tool({:running, {name, cmd, opts}, task}) do
-    Printer.info([:magenta, "=> running ", :bright, to_string(name)])
+    Printer.info([:magenta, "=> running "] ++ format_tool_name(name))
     Printer.info()
     IO.write(IO.ANSI.faint())
 
@@ -212,7 +136,7 @@ defmodule ExCheck.Check do
 
   defp reprint_errors(failed_tools) do
     Enum.each(failed_tools, fn {_, {name, _, _}, {_, output, _}} ->
-      Printer.info([:red, "=> reprinting errors from ", :bright, to_string(name)])
+      Printer.info([:red, "=> reprinting errors from "] ++ format_tool_name(name))
       Printer.info()
       IO.write(output)
       if output_needs_padding?(output), do: Printer.info()
@@ -224,31 +148,69 @@ defmodule ExCheck.Check do
     Printer.info()
 
     items
-    |> Enum.sort()
+    |> Enum.sort_by(&get_summary_item_order/1)
     |> Enum.each(&print_summary_item(&1, opts))
 
     Printer.info()
   end
 
+  defp get_summary_item_order({:ok, {name, _, _}, _}), do: {0, normalize_tool_name(name)}
+  defp get_summary_item_order({:error, {name, _, _}, _}), do: {1, normalize_tool_name(name)}
+  defp get_summary_item_order({:skipped, name, _}), do: {2, normalize_tool_name(name)}
+
+  defp normalize_tool_name(name = {_, _}), do: name
+  defp normalize_tool_name(name), do: {name, 0}
+
   defp print_summary_item({:ok, {name, _, _}, {_, _, duration}}, _) do
+    name = format_tool_name(name)
     took = format_duration(duration)
-    name = to_string(name)
-    Printer.info([:green, " ✓ ", :bright, name, :normal, " success in ", :bright, took])
+    Printer.info([:green, " ✓ ", name, " success in ", b(took)])
   end
 
   defp print_summary_item({:error, {name, _, _}, {code, _, duration}}, _) do
-    n = :normal
-    b = :bright
-    name = to_string(name)
-    code_string = to_string(code)
+    name = format_tool_name(name)
     took = format_duration(duration)
-    Printer.info([:red, " ✕ ", b, name, n, " error code ", b, code_string, n, " in ", b, took])
+    Printer.info([:red, " ✕ ", name, " error code ", b(code), " in ", b(took)])
   end
 
   defp print_summary_item({:skipped, name, reason}, opts) do
     if Keyword.get(opts, :skipped, true) do
-      Printer.info([:cyan, "   ", :bright, to_string(name), :normal, " skipped due to "] ++ reason)
+      name = format_tool_name(name)
+      reason = format_skip_reason(reason)
+      Printer.info([:cyan, "   ", name, " skipped due to ", reason])
     end
+  end
+
+  defp format_skip_reason({:run_after, name}) do
+    ["broken tool dependency ", format_tool_name(name)]
+  end
+
+  defp format_skip_reason({:package, name}) do
+    ["missing package ", b(name)]
+  end
+
+  defp format_skip_reason({:package, name, app}) do
+    ["missing package ", b(name), " in ", b(app)]
+  end
+
+  defp format_skip_reason({:file, name}) do
+    ["missing file ", b(name)]
+  end
+
+  defp format_skip_reason({:cd, cd}) do
+    ["missing directory ", b(cd)]
+  end
+
+  defp format_tool_name(name) when is_atom(name) do
+    b(name)
+  end
+
+  defp format_tool_name({name, app}) when is_atom(name) do
+    [b(name), " in ", b(app)]
+  end
+
+  defp b(inner) do
+    [:bright, to_string(inner), :normal]
   end
 
   defp format_duration(secs) do