+++ lib/mix/tasks/check.ex
@@ -27,6 +27,9 @@ defmodule Mix.Tasks.Check do
   - [`:ex_doc`] - compiles the project documentation in order to ensure that there are no issues
     that would make it impossible for docs to get collected and assembled
 
+  - [`:npm_test`] - runs JavaScript tests in projects with front-end assets embedded in `assets`
+    directory and `package.json` in it (default for Phoenix apps)
+
   You can disable or adjust curated tools as well as add custom ones via the configuration file.
 
   ## Workflow
@@ -86,6 +89,28 @@ defmodule Mix.Tasks.Check do
   they'll be skipped if their dependencies won't be run at all e.g. due to using `--except` command
   line option or a missing/circular dependency.
 
+  ### Umbrella projects
+
+  Task comes with extensive support for umbrella projects. The most notable feature is the ability
+  to run tools recursively for each child app separately. It's similar to flagging Mix tasks as
+  recursive but empowered with following extra benefits:
+
+  - runs recursively not just Mix tasks, but also arbitrary scripts & commands
+  - runs tools on child apps in parallel
+  - allows tools to target only specific child apps
+  - presents failures & run durations for each child app separately
+  - detects if curated tools should run for each child app separately
+  - builds separate cross-tool dependency chains for each child app
+
+  You may want to disable parallel execution of the tool on child apps (`parallel: false` under
+  `:umbrella` tool option) if it uses the same resources across tool runs against different child
+  apps. An example of that could be `ex_unit` that, depending on a project and test dependencies,
+  may involve mutating the same database in test suites belonging to separate child apps.
+
+  You may have the tool run *only* at the root level of the umbrella by disabling the recursive
+  execution (`recursive: false` under `:umbrella` tool option) and targeting an empty list of child
+  apps (`apps: []` under `:umbrella` tool option).
+
   ## Configuration file
 
   Check configuration may be adjusted with the optional `.check.exs` file. Task will load the
@@ -93,8 +118,7 @@ defmodule Mix.Tasks.Check do
 
   1. Default stock configuration.
   2. `.check.exs` in user home directory.
-  3. `.check.exs` in umbrella root directory when called from sub-project.
-  4. `.check.exs` in current project directory.
+  3. `.check.exs` in current project directory (or umbrella root for an umbrella project).
 
   Configuration file should evaluate to keyword list with following options:
 
@@ -111,13 +135,20 @@ defmodule Mix.Tasks.Check do
   - `:order` - integer that controls the order in which tool output is presented (default: `0`)
   - `:run_after` - list of tool names (atoms) as deps that must finish running before tool start
   - `:enable_ansi` - toggles extending Elixir/Mix commands to have ANSI enabled (default: `true`)
-  - `:require_deps` - list of package names (atoms) that must be present or tool will be skipped
-  - `:require_files` - list of filenames (strings) that must be present or tool will be skipped
+  - `:umbrella` - configures the tool behaviour in an umbrella project
+
+  Umbrella configuration under `:umbrella` key is a keyword list with following options:
+
+  - `:recursive` - toggles running the tool on each child app separately as opposed to running it
+    once from umbrella root (default: `true` except for non-recursive Mix tasks)
+  - `:parallel` - toggles running tool in parallel on all child apps (default: `true`)
+  - `:apps` - list of umbrella child app names targeted by the tool (default: all apps)
 
   You may also use one of the shorthand tool tuple forms:
 
-  - `{:tool_name, enabled}` where `enabled` is a boolean that translates into the `:enabled` option
-  - `{:tool_name, command}` where `command` is a binary that translates into the `:command` option
+  - `{:tool_name, enabled}` where `enabled` corresponds to the `:enabled` option
+  - `{:tool_name, command}` where `command` corresponds to the `:command` option
+  - `{:tool_name, command, opts}` where `command` corresponds to the `:command` option
 
   Use the `mix check.gen.config` task to generate sample configuration that comes with well-commented examples to help you get started.
 
@@ -135,6 +166,7 @@ defmodule Mix.Tasks.Check do
   [`:sobelow`]: https://hexdocs.pm/sobelow
   [`:dialyzer`]: https://hexdocs.pm/dialyxir
   [`:ex_doc`]: https://hexdocs.pm/ex_doc
+  [`:npm_test`]: https://docs.npmjs.com/cli/test.html
   """
 
   use Mix.Task