+++ test/ex_check/umbrella_project_cases/non_recursive_test.exs
@@ -1,0 +1,22 @@
+defmodule ExCheck.UmbrellaProjectCases.NonRecursiveTest do
+  use ExCheck.UmbrellaProjectCase, async: true
+
+  @config """
+  [
+    tools: [
+      {:ex_unit, umbrella: [recursive: false], detect: []},
+    ]
+  ]
+  """
+
+  test "non-recursive", %{project_dirs: [project_root_dir | _]} do
+    config_path = Path.join(project_root_dir, ".check.exs")
+    File.write!(config_path, @config)
+
+    assert {output, 0} = System.cmd("mix", ~w[check], cd: project_root_dir)
+
+    assert String.contains?(output, "ex_unit success")
+    refute String.contains?(output, "ex_unit in child_a success")
+    refute String.contains?(output, "ex_unit in child_b success")
+  end
+end